requirements.txt
aerich
aiofiles
aioredis
aiosqlite
asgiref
async-timeout
asyncpg
babel
bcrypt
cffi
click
ddlparse
dictdiffer
fastapi-admin
fastapi
h11
hiredis
httptools
markupsafe
pendulum
pycparser
pydantic
pyjwt
pyparsing
pypika-tortoise
python-dateutil
python-dotenv
python-multipart
pytz
pytzdata
pyyaml
six
starlette
tortoise-orm
typing-extensions
uvicorn[standard]
uvloop
watchgod
websockets
aiofile
grpcio
grpcio-tools
